add query.d.ts
