+++ app/frontend/src/types/index.d.ts
@@ -38,6 +38,7 @@ export * from './advancedSearchConditions';
 // ============================================
 // Condition Types
 // ============================================
+export * from './query';
 export * from './condition';
 export * from './conditionRestore';
 // ============================================
